Schools on Client side
Although the PHP has a long way to go, JavaScript and HTML and CSS are
taking their shape. Its is a whole new site, the main site of course,
but you can also see now quite clearly what schools is going to be like.
This is the code for Preview 01 on the Edunet Youtube and Facebook
channels.